no subir

+++ src/pages/Users/DeleteUser.jsx
@@ -3,7 +3,7 @@ import deleteUserService from '../../services/deleteUserService';
 
 
 const DeleteUser = ({ userId, onClose, onDelete}) => {
-    
+    const [showSuccess, setShowSuccess] = useState(false); // Estado para manejar el mensaje de eliminado con exito.
     // Función que maneja la eliminación del usuario
     const handleDeleteUser = async () => {
 
@@ -14,9 +14,16 @@ const DeleteUser = ({ userId, onClose, onDelete}) => {
         }
 
         try {
-            const deleteUser = await deleteUserService(userId); // Llama al servicio para eliminar el usuario y espera la respuesta
-            onDelete(); // Llama a onDelete después de la eliminación
-            onClose(); // Cierra el modal
+            const deleteRegister = await deleteUserService(userId); // Llama al servicio para eliminar el usuario y espera la respuesta
+            if (deleteRegister) {
+                setShowSuccess(true); // Muestra la alerta
+                setTimeout(() => {
+                    setShowSuccess(false); // Oculta la alerta después de 2 segundos
+                    onDelete(); // Llama a onDelete después de la eliminación
+                    onClose(); // Cierra el modal
+                },700); 
+            }
+            
         } catch (error) {
             console.error('Error al eliminar usuario:', error);
         }
@@ -28,10 +35,14 @@ const DeleteUser = ({ userId, onClose, onDelete}) => {
                 <Modal.Title>Eliminar Usuario</Modal.Title>
             </Modal.Header>
             <Modal.Body>
+                {showSuccess && (
+                    <Alert variant="success" className="custom-alert">
+                        Registro eliminado con éxito.
+                    </Alert>
+                )}
                 <p>
                     ¿Estas seguro de eliminar este Usuario?
                 </p>
-                 
             </Modal.Body>
             <Modal.Footer>
                 {/* Botón para eliminar usuario */}
